# openapi/openapi.yaml
openapi: 3.0.4
info:
  title: Tourism Monitoring API
  description: API для мониторинга и анализа туристических маршрутов в Москве
  version: "1.0.0"
  contact:
    name: rightDesire
servers:
  - url: http://localhost:8080/api
    description: Dev server
...
paths:
  /users/register:
    post:
      tags:
        - user
      summary: Регистрация нового пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegisterRequest"
      responses:
        "201":
          description: Пользователь успешно зарегистрирован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDTO"
        "400":
          description: Некорректный ввод
  /users/login:
    post:
      tags:
        - user
      summary: Аутентификация пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLoginRequest"
      responses:
        "200":
          description: Успешный вход в систему
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenDTO"
        "401":
          description: Неавторизовано
  /users/{userId}:
    get:
      tags:
        - user
      summary: Получить профиль пользователя
      parameters:
        - $ref: "#/components/parameters/userId"
      responses:
        "200":
          description: Профиль пользователя получен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDTO"
        "404":
          description: Пользователь не найден
    put:
      tags:
        - user
      summary: Обновление профиля пользователя
      parameters:
        - $ref: "#/components/parameters/userId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdateRequest"
      responses:
        "200":
          description: Пользователь успешно обновлен
        "400":
          description: Некорректный ввод
...
components:
  parameters:
    userId:
      name: userId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    UserRegisterRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        fullName:
          type: string
      required:
        - email
        - password
    UserLoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthTokenDTO:
      type: object
      properties:
        token:
          type: string
    UserDTO:
      type: object
      properties:
        uuid:
          type: string
        email:
          type: string
        fullName:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserUpdateRequest:
      type: object
      properties:
        email:
          type: string
        fullName:
          type: string
        password:
          type: string
...